Update incremental build perf tests
