+++ build-tools/automation/yaml-templates/run-integrated-regression-tests.yaml
@@ -90,13 +90,6 @@ steps:
   parameters:
     nunitConsole: $(System.DefaultWorkingDirectory)/NUnit.ConsoleRunner/tools/nunit3-console.exe
     testAssembly: $(System.DefaultWorkingDirectory)/XQA.Android/XQA.Android.dll
-    testRunTitle: Incremental Build Tests
-    nunitConsoleExtraArgs: --where "cat == BuildPerformance"
-
-- template: run-nunit-tests.yaml
-  parameters:
-    nunitConsole: $(System.DefaultWorkingDirectory)/NUnit.ConsoleRunner/tools/nunit3-console.exe
-    testAssembly: $(System.DefaultWorkingDirectory)/XQA.Android/XQA.Android.dll
     testRunTitle: AOT Tests
     nunitConsoleExtraArgs: --where "cat == AotSupport"
